refactored out gif and meme functions into another file
